refactor: establish layer boundaries and extract the dashboard contract (#600)
The audit found the dependency arrows pointing the wrong way. The dashboard's
data contract was declared inside `DashboardClient.tsx` — a 1,200-line
`'use client'` component — so `lib/dashboardOverview.ts`, the PDF report, the
nav badge and the report route's tests each imported a type out of the UI. A
dashboard edit could break the server, the report and the navigation, and there
was no rule saying where a type, hook or business rule belonged.

This is step 1 of the issue's rollout — contracts and shared UI hooks, no
behavior change:

- `features/dashboard/contracts.ts` now owns the overview DTOs (DashboardData,
  GoalData, InsuranceData, FundBreakdownItem, NonFundUnallocatedItem) plus the
  InvRow/InvTranche row model. Layer-neutral by construction: no React, no
  `'use client'`, no imports.
- `maturityCardItems.ts` — pure grouping logic shared by the dashboard card and
  the nav badge — moves to `features/dashboard/`, so navigation no longer
  reaches into `app/assets/`.
- `useDialogA11y` / `useDialogMount` move out of Planning into
  `components/ui/`. Five other features imported them from there.
- `docs/architecture.md` states the one rule: `app/` is the top layer, the arrow
  only points up, and where each kind of file goes.
- `__tests__/architecture.test.ts` enforces it — an import from `app/` in any
  lower layer, a route handler or the nav chrome importing feature UI, or a type
  exported from `DashboardClient` fails the build. It was written first and was
  red on four counts.

Coverage floors ratchet up (60.88/57.91/59.38/64.22 → 61.69/58.20/60.42/65.14):
moving fully-covered code out of files the report counts as largely untested
lifted the repo numbers.

Left for follow-ups, deliberately incremental: splitting `app/assets/` into
domain feature folders, folding `app/components/` into the top-level root, and
breaking up the oversized Dashboard / Maturity Resolve / Fund Library modules.
