+++ app/assets/components/DesktopInsuranceDetail.tsx
@@ -7,7 +7,7 @@ import { toast } from 'sonner'
 import { fmt, fmtCompact } from '@/lib/formatters'
 import { formatIntVN, parseIntVN } from '@/lib/numberFormat'
 import { STATUS_COLOR, BAR_COLOR_DETAIL, COVERAGE_OPTIONS, insurancePaidState, insuranceStatusLabel } from './insuranceShared'
-import type { InsuranceData } from '../DashboardClient'
+import type { InsuranceData } from '@/features/dashboard/contracts'
 import LogInsurancePaymentModal from './LogInsurancePaymentModal'
 import { TxRowsSkeleton } from './Skeletons'
 import LoadError from './LoadError'
