+++ app/assets/components/GoalDetailSheet.tsx
@@ -10,7 +10,7 @@ import { useLocale, useTranslations } from 'next-intl'
 import { toast } from 'sonner'
 import { fmt, fmtCompact, fmtPct } from '@/lib/formatters'
 import { formatIntVN, parseIntVN } from '@/lib/numberFormat'
-import type { GoalData, FundBreakdownItem } from '../DashboardClient'
+import type { GoalData, FundBreakdownItem } from '@/features/dashboard/contracts'
 import TransactionHistorySheet, { type PurchaseHistoryRow } from './TransactionHistorySheet'
 import { SellWithdrawSheet } from './SellWithdrawSheet'
 import { MaturityResolveSheet } from './MaturityResolveSheet'
