+++ app/assets/components/__tests__/DesktopGoalDetail.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { render, screen, waitFor, fireEvent } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import DesktopGoalDetail from '../DesktopGoalDetail'
-import type { GoalData } from '../../DashboardClient'
+import type { GoalData } from '@/features/dashboard/contracts'
 
 const { toastErrorMock } = vi.hoisted(() => ({ toastErrorMock: vi.fn() }))
 vi.mock('sonner', () => ({ toast: { error: toastErrorMock, success: vi.fn() } }))
