+++ app/assets/components/__tests__/DesktopInsuranceList.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import DesktopInsuranceList from '../DesktopInsuranceList'
-import type { InsuranceData } from '../../DashboardClient'
+import type { InsuranceData } from '@/features/dashboard/contracts'
 
 const KEY = 'cairn.insuranceCoachDismissed'
