+++ app/assets/components/__tests__/DesktopNetWorthPanel.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from 'vitest'
 import { render, screen, fireEvent } from '@testing-library/react'
 import DesktopNetWorthPanel from '../DesktopNetWorthPanel'
-import type { DashboardData } from '../../DashboardClient'
+import type { DashboardData } from '@/features/dashboard/contracts'
 import type { AllocationTotals } from '../../overviewData'
 
 // DesktopNetWorthPanel takes data/allocationTotals/locale as props (no next-intl),
