+++ app/assets/components/__tests__/GoalDetailSheet.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import GoalDetailSheet from '../GoalDetailSheet'
-import type { GoalData } from '../../DashboardClient'
+import type { GoalData } from '@/features/dashboard/contracts'
 
 vi.mock('next-intl', () => ({
   useLocale: () => 'en',
